requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
requests
django-cors-headers